requirements.txt
# Core dependencies
transformers>=4.34.0
accelerate>=0.23.0
bitsandbytes>=0.41.1
torch>=2.0.1
sentence-transformers>=2.2.2

# Document processing
langchain>=0.0.312
langchain-community>=0.0.10
pypdf>=3.15.1
docx2txt>=0.8
Pillow>=10.0.0
pytesseract>=0.3.10
unstructured>=0.10.30
pdf2image>=1.16.3

# Vector database
faiss-cpu>=1.7.4
chromadb>=0.4.18

# LLM integration
huggingface_hub>=0.17.3
einops>=0.7.0
ctransformers>=0.2.27

# Utilities
tqdm>=4.66.1
pydantic>=2.4.2
python-dotenv>=1.0.0 